[CHORE]: run reconciliation in batches
